added retrive_sensors_data to rows_detector_simulator
